move day header out
